requirements.txt
alembic==1.12.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pytest==7.4.3
pytest-html==4.1.1
pytest-metadata==3.0.0
python-dotenv==1.0.0
redis==5.0.1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.1
